+++ src/transformation/visitors/language-extensions/vararg.ts
@@ -1,6 +1,14 @@
 import * as ts from "typescript";
 import { TransformationContext } from "../../context";
 import * as extensions from "../../utils/language-extensions";
+import { Scope, ScopeType } from "../../utils/scope";
+
+export function isGlobalVarargConstant(context: TransformationContext, symbol: ts.Symbol, scope: Scope) {
+    return (
+        scope.type === ScopeType.File &&
+        extensions.isExtensionValue(context, symbol, extensions.ExtensionKind.VarargConstant)
+    );
+}
 
 export function isVarargConstantNode(context: TransformationContext, node: ts.Node): boolean {
     const symbol = context.checker.getSymbolAtLocation(node);
